crdyboot: Change PartitionIo to not rely on a partition-specific protocol
As seen in b/388506108, relying on protocols for partitions other than
the ESP can fail on some firmware. Change PartitionIo so that it uses a
`ScopedDiskIo` for the entire disk, rather than an individual
partition. PartitionIo now has a `byte_range` field indicating what part
of the disk contains the partition, and the read/write methods are
updated accordingly.

In the tests, `read_disk` is changed to expect the whole-disk Hd1 handle
instead of Hd1State.

This commit makes `find_partition_by_name` unused, but cleaning that up
(along with some tests) is deferred to a later commit to keep this one
smaller.

BUG=b:409770922
TEST=cargo xtask check
TEST=test boot on x1cg6

Change-Id: Ide082d38bb521a3f001e5867633d259efd97ebda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6448617
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>