xtask: Add test to ensure PartitionInfo is never used
In b/397698913, we removed use of PartitionInfo because it's not
supported on all devices. This was actually the second time we removed
use of PartitionInfo for that reason; the previous time was in 2021, but
we later forgot and started using it again.

Add code in xtask to check for banned patterns, and use that to ensure
that PartitionInfo is never used again.

BUG=none
TEST=cargo xtask check

Change-Id: I6de246df23217e6f8cdb71dc866009db40564831
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6482460
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>