crdyboot: Don't delete malformed fwupd variables
This wasn't implemented correctly, in that the code was conditionally
deleting variables while iterating through variable keys, which the spec
says not to do.

While this could be fixed to not delete during iteration, there's really
no strong reason for this deletion operation to occur; the behavior was
simply copied from fwupd-efi. Runtime-accessible fwupd variables will be
deleted by fwupd when the OS boots [1]. Non-runtime-accessible malformed
fwupd variables could theoretically be created, but there's no
particular reason to think this would happen, so not worth having code
to handle it.

[1]: https://github.com/fwupd/fwupd/blob/f2a89e48f746b940d1d8e917e7f103939d23898f/plugins/uefi-capsule/fu-uefi-capsule-plugin.c#L1202

BUG=b:403635944
TEST=cargo xtask check

Change-Id: Idf386d8489ff8140d012c3d0f8d962b0d31e4346
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6604629
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Nehemiah Dureus <ndureus@chromium.org>