crdyboot: Fix flexor on the HP Stream 11 Pro G2
Add another case to connect_disk_handles. This is slightly different
from the other cases, since there was no clear disk protocol to key off
of on this device.

It's possible that this might be a better general fix. As noted in the
TODO comment, the protocol-specific cases might no longer be needed, but
more testing will be needed before they can be removed.

BUG=b:414622656
TEST=cargo xtask check
TEST=Boot flexor on the HP Stream 11 Pro G2

Change-Id: I852a28d1300197e1f358778d15f47f48e71c1851
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6665078
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>