boot_message: Add BootloaderMessage struct
Add an android bootloader message struct for parsing
the recovery bootloader message struct from a buffer.

Parse the command member as it will be the part of
the struct that is checked.

Use the existing version of the zerocopy crate in use
by the bootimg dependency.

BUG=b:409535565
TEST=cargo xtask check

Change-Id: I5b148942858d608aa22e108c95195665e5e3caaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6951567
Reviewed-by: Allen Ballway <ballway@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>