avb: Move boot image parsing to its own module
Reorganize this code so the parsing of the boot partition
data is clearly separated from the use of the results.
This should make it easier to refactor and update the
code.
This is just a move of the current code with the
necessary visibility no other refactors are made.

BUG=b:444242354
TEST=cargo xtask check
TEST=confirm image boots to the kernel in a vm

Change-Id: Ic77e43485cf4ab85fead9c48994dacee33e4c91c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6933729
Tested-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Allen Ballway <ballway@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>