crdyboot: Rework InitramfsProtoHolder.proto_ptr
Use `&raw const` instead of `ptr::from_ref` to avoid creating an
intermediate reference. It probably doesn't matter in this case, but
avoids the subtleties of mixing references and raw pointers.

Also add an intermediate variable with an explicit type to make it
clearer that we start with the correct pointer type (i.e. a pointer to
the contents of the box, not accidentally a pointer to a Box itself).

BUG=none
TEST=cargo xtask check

Change-Id: I4d58a4f5220dc51bd27314e9dcf044cd5e640784
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/7018990
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>